+++ dmr/openapi/openapi.py
@@ -5,32 +5,29 @@ from dmr.openapi.mappers.schema_normalization import (
     DumpedSchema,
     dump_schema,
 )
+from dmr.openapi.objects.components import Components
+from dmr.openapi.objects.external_documentation import ExternalDocumentation
+from dmr.openapi.objects.info import Info
+from dmr.openapi.objects.path_item import PathItem
+from dmr.openapi.objects.paths import Paths
+from dmr.openapi.objects.reference import Reference
+from dmr.openapi.objects.security_requirement import SecurityRequirement
+from dmr.openapi.objects.server import Server
+from dmr.openapi.objects.tag import Tag
 
 if TYPE_CHECKING:
     from jsonschema_path.typing import Schema
     from openapi_spec_validator.validation.types import SpecValidatorType
 
-    from dmr.openapi.objects.components import Components
-    from dmr.openapi.objects.external_documentation import ExternalDocumentation
-    from dmr.openapi.objects.info import Info
-    from dmr.openapi.objects.path_item import PathItem
-    from dmr.openapi.objects.paths import Paths
-    from dmr.openapi.objects.reference import Reference
-    from dmr.openapi.objects.security_requirement import SecurityRequirement
-    from dmr.openapi.objects.server import Server
-    from dmr.openapi.objects.tag import Tag
+    class _ValidateSpecProto(Protocol):
+        def __call__(
+            self,
+            spec: 'Schema',
+            base_uri: str = '',
+            cls: 'SpecValidatorType | None' = None,  # noqa: WPS117
+        ) -> None: ...
 
-
-class _ValidateSpecProto(Protocol):
-    def __call__(
-        self,
-        spec: 'Schema',
-        base_uri: str = '',
-        cls: 'SpecValidatorType | None' = None,  # noqa: WPS117
-    ) -> None: ...
-
-
-_validate_spec: _ValidateSpecProto | None
+    _validate_spec: _ValidateSpecProto | None
 
 try:
     # There's a mismatch of checks with mypyc and mypy,
@@ -44,18 +41,25 @@ except ImportError:  # pragma: no cover
 
 @dataclasses.dataclass(kw_only=True, slots=True)
 class OpenAPI:
-    """This is the root object of the OpenAPI document."""
+    """
+    This is the root object of the OpenAPI document.
 
-    info: 'Info'
+    .. versionchanged:: 0.13.0
+        Moved from ``dmr.openapi.objects.OpenAPI``
+        to ``dmr.openapi.openapi.OpenAPI``.
+
+    """
+
+    info: Info  # noqa: WPS110
     openapi: str
     json_schema_dialect: str | None = None
-    servers: list['Server'] | None = None
-    paths: 'Paths | None' = None
-    webhooks: dict[str, 'PathItem | Reference'] | None = None
-    components: 'Components | None' = None
-    security: list['SecurityRequirement'] | None = None
-    tags: list['Tag'] | None = None
-    external_docs: 'ExternalDocumentation | None' = None
+    servers: list[Server] | None = None
+    paths: Paths | None = None
+    webhooks: dict[str, PathItem | Reference] | None = None
+    components: Components | None = None
+    security: list[SecurityRequirement] | None = None
+    tags: list[Tag] | None = None
+    external_docs: ExternalDocumentation | None = None
 
     _validated: bool = dataclasses.field(
         default=False,